% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{get_mapping_table}
\alias{get_mapping_table}
\title{Get the mapping dataset for specified species}
\usage{
get_mapping_table(from_species, to_species, version = c("es96"))
}
\arguments{
\item{from_species}{From which species}

\item{to_species}{To which species}

\item{version}{Ensembl version}
}
\value{
A data.table that has all the gene mappings for specied parameters
}
\description{
Get the mapping dataset for specified species
}
