% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{map_genes}
\alias{map_genes}
\title{Map gene identifiers within or cross species. Type \code{genemap::SPECIES} to see the full list of supported species}
\usage{
map_genes(from_species = SPECIES, from_id_list,
  from_id_type = c("ensemblgid", "ncbigid", "symbol"),
  to_species = SPECIES, to_id_type = c("ensemblgid", "ncbigid",
  "symbol"))
}
\arguments{
\item{from_species}{From which species}

\item{from_id_list}{Gene identifiers to be mapped}

\item{from_id_type}{The type of the identifiers in \code{from_id_list}}

\item{to_species}{To which species}

\item{to_id_type}{The type of the identifiers to map to}
}
\value{
A named list of the same size with \code{from_id_list}. Each gene from \code{from_id_list} will be mapped to an element in this result list.
    \code{$mapped_genes} will contain the mapped identifiers and  \code{$mapped_gene_symbols} will contain corresponding symbols of these mapped genes.
    \code{$gene_symbol} will be the corresponding gene symbol of the input gene identifier. Name of each element is the original gene identifier used to map.
}
\description{
Map gene identifiers within or cross species. Type \code{genemap::SPECIES} to see the full list of supported species
}
\examples{
# within species mapping
map_genes(
  from_species = "hs",
  from_id_list = c("ENPP4", "GCLC"),
  from_id_type = "symbol",
  to_species = "hs"
)
# cross species mapping
map_genes(
  from_species = "hs",
  from_id_list = c("ENPP4", "GCLC"),
  from_id_type = "symbol",
  to_species = "mm",
  to_id_type = "symbol"
)
}
