% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{map_genes}
\alias{map_genes}
\title{Map gene identifiers within or cross species}
\usage{
map_genes(from_species = SPECIES, from_id_list,
  from_id_type = c("ensemblgid", "ncbigid", "symbol"),
  to_species = SPECIES, to_id_type = c("ensemblgid", "ncbigid",
  "symbol"))
}
\arguments{
\item{from_species}{from which species}

\item{from_id_list}{gene identifiers to be mapped}

\item{from_id_type}{the type of the identifiers in \code{from_id_list}}

\item{to_species}{map to which species}

\item{to_id_type}{the type of the identifiers to map to}
}
\value{
The mapped gene identifiers
}
\description{
Map gene identifiers within or cross species
}
\examples{
map_genes(from_species = "hs", from_id_list = c("ENPP4", "GCLC"), from_id_type = "symbol", to_species = "hs") # within species mapping
map_genes(from_species = "hs", from_id_list = c("ENPP4", "GCLC"), from_id_type = "symbol", to_species = "mm", to_id_type = "symbol") # cross species mapping
}
